Added a plugin implementing the Heston process.
The Heston model simulates equity or index prices taking into account stochastic volatility effects. The main feature of the model is that the price process follows a geometric brownian motion with a stochastic volatility while the volatility follows a square root mean reverting process. Usually the correlation is negative, so that a lowering in the stock price is correlated with an increasing in the volatility. Once installed the plug-in offers the possibility of using two new processes, the Heston process and the Heston time dependent drift process and to calibrate them to a series of call prices.